move functions to funtions.h
